+++ CaseBased/bin/.gitignore
@@ -1,3 +1,4 @@
 /cbr/
-/model/
 /connector/
+/model/
+/similarity/
