--- CaseBased/data/treatments.csv
@@ -1,28 +1,0 @@
-#diagnosis;gender;age;treatment
-
-iron-deficiency_anemia;male;25;iron_pills,diet,blood_transfusion
-iron-deficiency_anemia;female;30;iron_pills,diet
-iron-deficiency_anemia;male;55;blood_transfusion
-iron-deficiency_anemia;female;60;blood_transfusion,diet
-iron-deficiency_anemia;male;15;iron_pills,diet
-diabetes_type_1;male;17;diet,excercise,insulin_injections,blood_sugar_monitoring
-diabetes_type_1;female;18;insulin_injections,excercise
-diabetes_type_1;male;27;diet,blood_sugar_monitoring,insulin_injections
-diabetes_type_1;female;52;insulin_injections,blood_sugar_monitoring
-diabetes_type_1;female;29;excercise
-diabetes_type_1;female;55;excercise
-diabetes_type_2;female;34;diet,insulin_injections,blood_sugar_monitoring,excercise
-diabetes_type_2;male;16;diet
-diabetes_type_2;female;50;excercise
-diabetes_type_2;male;72;blood_sugar_monitoring
-diabetes_type_2;female;31;insulin_injections
-diabetes_type_2;male;14;diet,excercise
-gestational_diabetes;female;28;diet
-gestational_diabetes;female;40;diet,blood_sugar_monitoring
-gestational_diabetes;female;20;diet,blood_sugar_monitoring,excercise
-gestational_diabetes;female;25;diet,excercise
-pernicious_anemia;male;43;vitamin_b12_injections
-pernicious_anemia;female;23;vitamin_b12_injections,blood_transfusion
-pernicious_anemia;male;51;blood_transfusion
-pernicious_anemia;female;41;vitamin_b12_injections,blood_transfusion
-
