+++ CaseBased/src/cbr/treatments/CbrApplication.java
@@ -2,11 +2,9 @@ package cbr.treatments;
 
 import java.util.Collection;
 
-import connector.treatments.CsvTreatmentConnector;
-import model.treatments.TreatmentsDescription;
+import connector.treatments.CsvConnector;
 import ucm.gaia.jcolibri.casebase.LinealCaseBase;
 import ucm.gaia.jcolibri.cbraplications.StandardCBRApplication;
-import ucm.gaia.jcolibri.cbrcore.Attribute;
 import ucm.gaia.jcolibri.cbrcore.CBRCase;
 import ucm.gaia.jcolibri.cbrcore.CBRCaseBase;
 import ucm.gaia.jcolibri.cbrcore.CBRQuery;
@@ -16,32 +14,23 @@ import ucm.gaia.jcolibri.method.retrieve.RetrievalResult;
 import ucm.gaia.jcolibri.method.retrieve.NNretrieval.NNConfig;
 import ucm.gaia.jcolibri.method.retrieve.NNretrieval.NNScoringMethod;
 import ucm.gaia.jcolibri.method.retrieve.NNretrieval.similarity.global.Average;
-import ucm.gaia.jcolibri.method.retrieve.NNretrieval.similarity.local.Equal;
-import ucm.gaia.jcolibri.method.retrieve.NNretrieval.similarity.local.Interval;
 import ucm.gaia.jcolibri.method.retrieve.selection.SelectCases;
 
-public class CbrTreatmentApplication implements StandardCBRApplication {
-
+public class CbrApplication implements StandardCBRApplication {
+	
 	Connector _connector;  /** Connector object */
 	CBRCaseBase _caseBase;  /** CaseBase object */
 
 	NNConfig simConfig;  /** KNN configuration */
-
+	
 	public void configure() throws ExecutionException {
-		_connector =  new CsvTreatmentConnector();
-
+		_connector =  new CsvConnector();
+		
 		_caseBase = new LinealCaseBase();  // Create a Lineal case base for in-memory organization
-
+		
 		simConfig = new NNConfig(); // KNN configuration
 		simConfig.setDescriptionSimFunction(new Average());  // global similarity function = average
-
-		simConfig.addMapping(new Attribute("diagnosis", TreatmentsDescription.class), new Equal());
-		simConfig.addMapping(new Attribute("gender", TreatmentsDescription.class), new Equal());
-		simConfig.addMapping(new Attribute("age", TreatmentsDescription.class), new Interval(100));
-		//simConfig.addMapping(new Attribute("treatment", TreatmentDescription.class), new Equal());
-
-		//simConfig.addMapping(new Attribute("treatment", TreatmentDescription.class), new ListTableSimilarity());
-
+		
 		// simConfig.addMapping(new Attribute("attribute", CaseDescription.class), new Interval(5));
 		// TODO
 
@@ -53,20 +42,19 @@ public class CbrTreatmentApplication implements StandardCBRApplication {
 		// EnumDistance - returns the similarity of two enum values as the their distance: sim(x,y) = |ord(x) - ord(y)|
 		// EnumCyclicDistance - computes the similarity between two enum values as their cyclic distance
 		// Table - uses a table to obtain the similarity between two values. Allowed values are Strings or Enums. The table is read from a text file.
-		// TableSimilarity(List<String> values).setSimilarity(value1,value2,similarity)
+		// TableSimilarity(List<String> values).setSimilarity(value1,value2,similarity) 
 	}
 
 	public void cycle(CBRQuery query) throws ExecutionException {
 		Collection<RetrievalResult> eval = NNScoringMethod.evaluateSimilarity(_caseBase.getCases(), query, simConfig);
-		eval = SelectCases.selectTopKRR(eval, 4);
+		eval = SelectCases.selectTopKRR(eval, 5);
 		System.out.println("Retrieved cases:");
 		for (RetrievalResult nse : eval)
 			System.out.println(nse.get_case().getDescription() + " -> " + nse.getEval());
-
 	}
 
 	public void postCycle() throws ExecutionException {
-
+		
 	}
 
 	public CBRCaseBase preCycle() throws ExecutionException {
@@ -75,29 +63,21 @@ public class CbrTreatmentApplication implements StandardCBRApplication {
 		for (CBRCase c: cases)
 			System.out.println(c.getDescription());
 		return _caseBase;
-
-
-
 	}
 
 	public static void main(String[] args) {
-		StandardCBRApplication recommender = new CbrTreatmentApplication();
+		StandardCBRApplication recommender = new CbrApplication();
 		try {
 			recommender.configure();
 
 			recommender.preCycle();
 
 			CBRQuery query = new CBRQuery();
-
-			TreatmentsDescription trDescription = new TreatmentsDescription();
-			trDescription.setDiagnosis("diabetes_type_1");
-			trDescription.setGender("female");
-			trDescription.setAge(60);
-
-
+			//CaseDescription caseDescription = new CaseDescription();
+			
 			// TODO
-
-			query.setDescription( trDescription );
+			
+			//query.setDescription( caseDescription );
 
 			recommender.cycle(query);
