--- CaseBased/src/connector/treatments/CsvTreatmentConnector.java
@@ -1,82 +1,0 @@
-package connector.treatments;
-
-import java.io.BufferedReader;
-import java.io.InputStreamReader;
-import java.net.URL;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.LinkedList;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
-
-import model.treatments.TreatmentsDescription;
-import ucm.gaia.jcolibri.cbrcore.CBRCase;
-import ucm.gaia.jcolibri.cbrcore.CaseBaseFilter;
-import ucm.gaia.jcolibri.cbrcore.Connector;
-import ucm.gaia.jcolibri.exception.InitializingException;
-import ucm.gaia.jcolibri.util.FileIO;
-
-
-public class CsvTreatmentConnector implements Connector {
-
-    @Override
-    public Collection<CBRCase> retrieveAllCases() {
-
-        LinkedList<CBRCase> cases = new LinkedList<CBRCase>();
-
-        try {
-            BufferedReader br = new BufferedReader(new InputStreamReader(FileIO.openFile("data/treatments.csv")));
-            if (br == null)
-                throw new Exception("Error opening file");
-
-            String line = "";
-            while ((line = br.readLine()) != null) {
-                if (line.startsWith("#") || (line.length() == 0))
-                    continue;
-                String[] values = line.split(";");
-
-                CBRCase cbrCase = new CBRCase();
-
-                TreatmentsDescription trDescription = new TreatmentsDescription();
-                trDescription.setDiagnosis(values[0]);
-                trDescription.setGender(values[1]);
-                trDescription.setAge(Integer.parseInt(values[2]));
-                trDescription.setTreatment(Arrays.asList(values[3].split(",")));
-
-                // TODO
-
-                cbrCase.setDescription(trDescription);
-                cases.add(cbrCase);
-            }
-            br.close();
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return cases;
-    }
-
-    @Override
-    public Collection<CBRCase> retrieveSomeCases(CaseBaseFilter arg0) {
-        return null;
-    }
-
-    @Override
-    public void storeCases(Collection<CBRCase> arg0) {
-    }
-
-    @Override
-    public void close() {
-    }
-
-    @Override
-    public void deleteCases(Collection<CBRCase> arg0) {
-    }
-
-    @Override
-    public void initFromXMLfile(URL arg0) throws InitializingException {
-    }
-
-
-}
