+++ CaseBased/src/model/treatments/Treatments.java
@@ -1,0 +1,19 @@
+package model.treatments;
+
+import ucm.gaia.jcolibri.cbrcore.Attribute;
+import ucm.gaia.jcolibri.cbrcore.CaseComponent;
+
+public class Treatments implements CaseComponent {
+
+
+	@Override
+	public String toString() {
+		return "";
+	}
+	
+	@Override
+	public Attribute getIdAttribute() {
+		return null;
+	}
+	
+}
